moving FontVal wrapper to a separate spec
(issue #2169)
