+++ tests/specifications/general_test.py
@@ -66,37 +66,6 @@ def test_check_036():
   assert "Failed to sanitize file!" in output
 
 
-def test_check_037():
-  """ MS Font Validator checks """
-  from fontbakery.specifications.general import com_google_fonts_check_037 as check
-
-  font = "data/test/mada/Mada-Regular.ttf"
-  RASTER_EXCEPTION_MESSAGE = ("MS-FonVal: An exception occurred"
-                              " during rasterization testing")
-  # we want to run all FValidator checks only once,
-  # so here we cache all results:
-  fval_results = list(check(font))
-
-  # Then we make sure that the freetype backend we're using
-  # supports the hinting instructions validation checks,
-  # which are refered to as "rasterization testing":
-  # (See also: https://github.com/googlefonts/fontbakery/issues/1524)
-  for status, message in fval_results:
-    assert RASTER_EXCEPTION_MESSAGE not in message
-
-    # and finally, we make sure that there wasn't an ERROR
-    # which would mean FontValidator is not properly installed:
-    assert status != ERROR
-
-  # Simulate FontVal missing.
-  old_path = os.environ["PATH"]
-  os.environ["PATH"] = ""
-  with pytest.raises(OSError) as _:
-    status, message = list(check(font))[-1]
-    assert status == ERROR
-  os.environ["PATH"] = old_path
-
-
 def NOT_IMPLEMENTED_test_check_038():
   """ FontForge validation outputs error messages? """
   # from fontbakery.specifications.general import com_google_fonts_check_038 as check
